requirements.txt
transformers==4.34.0
datasets==2.12.0
peft==0.4.0
pandas==2.0.2
numpy==1.24.3
ipdb==0.13.13
evaluate==0.4.0
scikit-learn==1.2.2
scipy==1.10.1
wandb==0.15.4
einops==0.6.1
bitsandbytes==0.39.0
nltk==3.8.1
accelerate==0.23.0
trl==0.4.7
rouge-score==0.1.2
absl-py==1.4.0
py7zr==0.20.5
tiktoken==0.4.0
ninja==1.11.1
packaging==23.1
sentencepiece==0.1.99
protobuf==3.20.*
ai21==1.2.7
openai==0.28.1
ujson==5.8.0


# for toolkit
pyyaml
ijson